# add look ahead for unpackTags