By doing two runs through the array, we can determine the exact
size of the map we should allocate ahead of time. In an added
benchmark, allocations per unpack and time per unpack are
reduced by about 20% across the board:
```
name          old time/op    new time/op    delta
UnpackTags-8    7.95µs ± 1%    6.29µs ± 7%  -20.90%  (p=0.029 n=4+4)

name          old alloc/op   new alloc/op   delta
UnpackTags-8    10.3kB ± 0%     7.9kB ± 0%  -23.22%  (p=0.029 n=4+4)

name          old allocs/op  new allocs/op  delta
UnpackTags-8      40.0 ± 0%      33.0 ± 0%  -17.50%  (p=0.029 n=4+4)
```