@thomersch 2021-01-22
Excellent stuff! Totally forgot that maps can be given a capacity.

The tests are failing, which seems to be caused by an indirect dependency, which requires Go 1.15, but gosmparse currently supports everything back to 1.6. All of that is not your fault, just a coincidence. I'll have a look at how to resolve that.