@thomersch 2021-01-22
Alright, gosmparse is now requiring Go 1.11 or higher, because those have module support.

Thank you for your work!